requirements.txt
langchain-core>=0.3.56
langchain-community>=0.3.12
langgraph>=0.3.25
langchain-mcp-adapters==0.0.9
mcp>=1.6.0
arize-phoenix>=8.28.1
opentelemetry-api>=1.32.1
opentelemetry-sdk>=1.32.1
opentelemetry-exporter-otlp>=1.32.1
oci>=2.154.0
oracledb>=3.1.1
rapidfuzz>=3.13.0
sentence-transformers>=2.2.2
numpy>=1.24.0
gradio>=5.0.0
